dec: add OrderedDecimal wrapper
Add an OrderedDecimal wrapper type, akin to ordered_float::OrderedFloat,
which provides an Ord and a Hash implementation for Decimal64 and
Decimal128. The order used sorts NaNs as greater than all other
elements and equal to themselves, but otherwise respects the PartialOrd
implementation of the underlying type. This is generally the ordering
that people want when they ask for a total ordering on a floating-point
type.

Co-authored-by: Nikhil Benesch <nikhil.benesch@gmail.com>