dec: reduce use of uninitialized arrays/vecs
Previously, we used uninitialized Decimal values for parsing.
This had the downside that the values' `lsu` array contained
uninitialized values after parsing because the underlying library
fills the `lsu` array with the parsed digits, but doesn't modify
the remaining elements, i.e. junk bytes remained in the `lsu`.

This commit instead uses Decimal values initialized to zero for
parsing, which zeroes out the `lsu` array.

This PR also initializes a arrays in a few spots, by the same
reasoning.