dec: refactor *_raw_parts to represent LSU values with u16s
The prior implementation of the raw_parts functions potentially
mangled endianness if ported across machines with different endianness.
This commit refactors this to make managing the LSU's conversion
to bytes an issue for the client.

This essentially undoes the majority of #42, which I had implemented
so as to provide an immediate LSU-to-byte transformation in the library,
rather than implementing the transformation in my application. That is
to say that #42 was over-eager and should not have been committed.