avoid `MaybeUninit`
The use of `MaybeUninit` was invalid in `quantum_matches`,
`partial_cmp`, `total_cmp`. The underlying decNumber library was *not*
guaranteed to zero-initialize the entire `Decimal` struct--in particular
it would often leave unneeded bytes in `lsu` uninitialized--but Rust
requires that the entire struct be initialized before calling
`assume_init`. So just zero initialize in Rust and avoid `MaybeUninit`
entirely.

Note that we do think that the use of `MaybeUninit` to avoid
zero-initializing the entire string formatting buffer in the
`fmt::Display` impls was valid, but to avoid a false positive with
@icmccorm's automatic verification tool, we're switching to zero
initialization there as well. The performance hit should be negligible,
and we can always switch back to `MaybeUninit` there in the future if it
becomes important.