dec: use decimal representation when converting to Decimal
This PR fixes a regression introduced in #83 which started serializing
floats using their scientific notation, which is much shorter than the
decimal notation and therefore faster.

It turns out that conversion from string to decimal is syntactic
instead of semantic. This means that the strings "12" and "1.2E2", which
are semantically the same number, parse into a different Decimal
instance.

This PR fixes the regression by serializing to the decimal
representation.

The overall improvement from 0.4.9 is much more mild as a result.

```
parse_decimal           time:   [2.4036 µs 2.4180 µs 2.4389 µs]
                        change: [-22.670% -21.979% -21.336%] (p = 0.00 < 0.05)
                        Performance has improved.
```

Signed-off-by: Petros Angelatos <petrosagg@gmail.com>